+++ libloragw/inc/loragw_hal.h
@@ -400,7 +400,7 @@ int lgw_abort_tx(void);
 @param trig_cnt_us pointer to receive timestamp value
 @return LGW_HAL_ERROR id the operation failed, LGW_HAL_SUCCESS else
 */
-int lgw_get_trigcnt(uint32_t* trig_cnt_us);
+int lgw_get_trigcnt(uint64_t* trig_cnt_us);
 
 /**
 @brief Allow user to check the version/options of the library once compiled
